ml/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.4
decorator==5.1.1
dynaconf==3.2.3
executing==2.0.0
fastapi==0.104.0
fastjsonschema==2.18.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.10.0
gluonnlp==0.10.0
graphviz==0.8.4
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
import-ipynb==0.1.4
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.4.0
keras==2.14.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
mxnet==1.6.0
mypy-extensions==1.0.0
nbformat==5.9.2
networkx==3.2
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.13
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.9
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
traitlets==5.12.0
transformers==4.34.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
wcwidth==0.2.8
